auth/requirements.txt
Flask
PyJWT
psycopg2-binary==2.9.9
